// protocol-designer/src/utils/labwareModuleCompatibility.d.ts
import { LabwareDefinition2, ModuleType } from '@opentrons/shared-data';
import { LabwareDefByDefURI } from '../labware-defs';
import { LabwareOnDeck } from '../step-forms';
export declare const COMPATIBLE_LABWARE_ALLOWLIST_BY_MODULE_TYPE: Record<ModuleType, Readonly<string[]>>;
export declare const getLabwareIsCompatible: (def: LabwareDefinition2, moduleType: ModuleType) => boolean;
export declare const ADAPTER_96_CHANNEL = "opentrons_flex_96_tiprack_adapter";
export declare const COMPATIBLE_LABWARE_ALLOWLIST_FOR_ADAPTER: Record<string, string[]>;
export declare const getLabwareCompatibleWithAdapter: (adapterLoadName?: string) => string[];
export declare const getLabwareIsCustom: (customLabwares: LabwareDefByDefURI, labwareOnDeck: LabwareOnDeck) => boolean;
export declare const getAdapterLabwareIsAMatch: (labwareId: string, allLabware: LabwareOnDeck[], draggedLabwareLoadname: string) => boolean;